requirements.txt
click==8.1.7
Werkzeug==3.0.1
fastapi==0.109.2
uvicorn[standard]==0.27.0
requests==2.31.0
pydantic==2.5.2
pydantic[email]==2.5.2
httpx==0.26.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
sqlalchemy==2.0.25
passlib==1.7.4
python-jose==3.3.0
python-multipart==0.0.6
matplotlib==3.8.2
pandas==2.1.4
bcrypt==4.1.2
aiohttp==3.9.1
aiosqlite==0.19.0
asyncpg==0.29.0
python-dateutil==2.8.2

# Development and testing
pytest==7.4.3
black==23.12.0
isort==5.13.2
mypy==1.8.0 